+++ src/math_utils.h
@@ -11,7 +11,7 @@ float    minf(float a, float b);
 uint16_t float_to_fixed(float v, float min, float max);
 float    fixed_to_float(uint16_t val, float min, float max);
 bool     is_in(float x, float a, float b);
-int      positive_mod(int x, int m);
+int16_t  positive_mod(int x, int m);
 
 // Vector algebra
 void  vec3_add(float result[3], const float a[3], const float b[3]);
